Rename `--use-pkcs11-spy` to `--pkcs11-spy-path` and make its value required.
clap could not parse the argv when the parameter was specified with no value,
because it would parse the subcommand name after the parameter as the value of
the parameter instead.

This commit also allows `--pkcs11-lib-path` and `--pkcs11-spy-path` to be set
via env vars, so that they don't have to be repeated for every invocation.